Started work on updating deprecated code
17 out of 18 tests passes
