+++ lib/main.js
@@ -4,22 +4,22 @@
 "use strict";
 
 //Import the APIs we need.
-const contextMenu = require("context-menu");
-const request = require("request");
-const selection = require("selection");
-const translate = require("translate");
-const self = require("self");
-const data = require("self").data;
-const tabs = require("tabs");
-const pageMod = require("page-mod");
-const localProcess = require("launch-local-process");
-const linkUtil = require("link-util");
-const osUtil = require("os-util");
-const widgetMain = require("widget-main");
-const ss = require("simple-storage");
+const contextMenu = require("sdk/context-menu");
+const request = require("sdk/request");
+const selection = require("sdk/selection");
+const translate = require("sdk/translate");
+const self = require("sdk/self");
+const data = require("sdk/self").data;
+const tabs = require("sdk/tabs");
+const pageMod = require("sdk/page-mod");
+const localProcess = require("sdk/launch-local-process");
+const linkUtil = require("sdk/link-util");
+const osUtil = require("sdk/os-util");
+const widgetMain = require("sdk/widget-main");
+const ss = require("sdk/simple-storage");
 const simple_prefs = require("sdk/simple-prefs").prefs;
 ////const jQuery = require("jquery-1.6.4-module.js").jQuery; // by alien
-const prefs = require("preferences-service");
+const prefs = require("sdk/preferences/service");
 
 var hyperlinkScanner = undefined; // PageMod
 var hyperlinkScannerWorkers = [];
