+++ lib/sdk/evallib.js
@@ -1,0 +1,7 @@
+/*
+ * License: www.mozilla.org/MPL/
+ */
+"use strict";
+
+// https://addons.mozilla.org/en-US/developers/docs/sdk/1.1/dev-guide/addon-development/commonjs.html
+exports.hello = function() { console.log("hallo from evallib.js"); };
