+++ lib/sdk/link-util.js
@@ -3,8 +3,8 @@
  */
 "use strict";
 
-const stringUtil = require("string-util");
-const osUtil = require("os-util");
+const stringUtil = require("sdk/string-util");
+const osUtil = require("sdk/os-util");
 
 //
 // used for Windows OS:
