+++ lib/sdk/string-util.js
