+++ lib/sdk/widget-main.js
@@ -7,11 +7,11 @@
      Creates the toolbar panel and the options panel
      and handles the events
  */
-const widget = require("widget");
-const data = require("self").data;
-const panel = require("panel");
-const tabs = require("tabs");
-const main = require("main");
+const widget = require("sdk/widget");
+const data = require("sdk/self").data;
+const panel = require("sdk/panel");
+const tabs = require("sdk/tabs");
+const main = require("../main");
 
 var onSettingsApplyCallback;
