+++ package.json
@@ -2,7 +2,7 @@
     "name": "alien-local-filesystem-links", 
     "license": "MPL 1.1/GPL 2.0/LGPL 2.1", 
     "author": "austrALIENsun", 
-    "version": "0.99.25", 
+    "version": "0.99.30", 
     "__comment-on-version": "/* australiensun: <major>.<minor>.<released for firefox version yyy> */", 
     "fullName": "Local Filesystem Links",
     "id": "jid1-JAzC7z53jemo5Q",
