+++ test/webserver/start_test_server.cmd
@@ -1,3 +1,3 @@
-set NODEJS="C:\Program Files (x86)\nodejs\node.exe"
+set NODEJS="C:\Program Files\nodejs\node.exe"
 %NODEJS% test-server.nodejs.js
 pause
